[REF] stock*: move test TestStockCommon2
move code common2.py in common.py and change the
name of class name to avoid confusing TestStockCommon (of common.py) vs
TestStockCommon (of common2.py)

task-2566396
